src/functions/pyproject.toml
[project]
name = "kb-ingestion-functions"
version = "0.1.0"
description = "Azure Functions for KB article ingestion — convert HTML to Markdown and index into AI Search"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-contentunderstanding>=1.0.0b1",
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.20.0",
    "azure-search-documents>=11.6.0b5",
    "azure-ai-inference>=1.0.0b1",
    "azure-functions>=1.21.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["fn_convert", "fn_index", "shared"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
